Move the ConditionedGmfComputer from the shakemap module to calc
Minor refactoring to move the ConditionedGmfComputer from the
hazardlib/shakemap module to hazardlib/calc, which is the more
appropriate location for the class, given the extent of code
shared with the regular GmfComputer.

The shakemap module instead is better suited to support the
different implementations of ShakeMap, including ones run by
the  Swiss Seismological Servicehazardlib/calc, which is the more
appropriate location for the class, given the extent of code
shared with the regular GmfComputer.

The shakemap module instead is better suited to support the
different implementations of ShakeMap, including ones run by
the Swiss Seismological Service (SED) in Switzerland,
INGV in Italy, Institute of Engineering Seismology and
Earthquake Engineering (ITSAK) in Greece, and ShakeMapEU.
